Fix ~600 decompiler artifacts in bulk
- GlobalUsings: add Extensions=TaleWorlds.Core, Path=System.IO, FaceGen=TaleWorlds.Core,
  GameMenuOption, and 9 more game-nested-type outer classes
- OnConditionDelegate type aliases in menu files (fix CS1503)
- JToken.op_Implicit( → (JToken)( in 3 API files (238 errors)
- _002Ector struct constructor calls → new TYPE(args) (112 errors)
- _003F → Vec2 in PlayerReinforcementMissionLogic
- AgentFlag bitwise: (AgentFlag)(x|n) → (AgentFlag)((int)x|n)
- (BaseType)this).Method() → base.Method() (CS1540)
- Missing out keywords in AiHelper calls (CS1620)
- SeasonalDiseaseSystem switch on WeatherEvent enum → cast to int
- BattleSideEnum bool-cast → (BattleSideEnum)(1-(int)x)
- KillCharacterAction alias in files with Diseases namespace ambiguity
- .ToHashSet() → new HashSet<string>(...) polyfill
- Add Microsoft.CSharp package for dynamic keyword
- Add System.Index/Range polyfill struct for net472

+++ src/AIInfluence.Diseases/SeasonalDiseaseSystem.cs
@@ -167,7 +167,7 @@ public static class SeasonalDiseaseSystem
 			WeatherEvent weatherEventInPosition = val2.GetWeatherEventInPosition(position2D);
 			WeatherEvent val3 = weatherEventInPosition;
 			WeatherEvent val4 = val3;
-			switch (val4 - 1)
+			switch ((int)val4 - 1)
 			{
 			case 0:
 				num += WeatherLightRainAdd;
